Removes all “nudges” from “x-element”.
There’s been a lot of discussion about the opinions “x-element” makes to
be helpful — and sometimes the accusation is that “x-element” is getting
in the way.

This PR removes _all_ such nudges. This should feel inline with our
choice to remove errors for attempting to access `internal` properties
on the host (e.g., `host.myInternalProp` just returns `undefined` now).

In particular, we no longer error for the following:

* If an attribute is set for an unserializable property — we no longer
  observe such attributes at all.

And, we no longer warn for the following:

* If you have a property that shadows the HTMLElement interface’s
  prototype chain — e.g., “title”, “draggable”, “role”, “dataset”,
  “style”, etc, etc. The assumption is that users of “x-element” are
  experts and they would only do this _on purpose_.
* If you have a property that shadows “data-*” attributes.
* If you have a property that shadows “aria-*” attributes.

Closes #77.