Allow special characters in quoted attributes.
Characters like “<”, “/”, and “>” are special in html and we use them
as boundary markers to determine where we are as we walk through the
result of a tagged template literal callback.

However, it’s perfectly allowable to use the characters in the markup
itself if they’re quoted. Take the following, resonable example:

```
html`<a href="https://example.com" target="${target}">click me</a>`
```

Previously, because of the “/” characters in the “href” attribute
declaration, the “${target}” binding would get missed. This is now
fixed to allow quoted, special characters.