Improve internal “template engine”.
Nearly all of the template engine details are kept internal — this
allowed us to ship a first pass of the code without needed high-levels
of scrutiny. However, we can (and will!) revisit this internal code to
make it more readable / self-documenting.

In addition, we can document some of the high-level concepts since it
may be helpful to understand the intent of some of the code.

Closes #187.