Use non-registered symbols as weak map keys.
This is now baseline as of 2023 so we ought to be able to use it. We use
symbols to prevent internal abstractions from being leaked into code
authored by integrators — for that, we just need an opaque token. But,
because weak maps previously needed to be keyed with an _object_, we
just created a dummy object as our key.

Now, we can take the more modern approach of keying with a symbol.

Closes #192.