Synchronously parse html / svg.
Previously, the parsing only happened after calling “render”. The issue
with this is that you lose the original line numbers on errors when
something goes wrong. By forcing the parse to happen immediately, you
can find the actual invocation of “html” within which the issue lies.

The only potential downside to this approach would be a performance hit
within our “update” path. This is because we would check whether the
“strings” from our tagged template function have been analyzed or not.
Checking for their existence in our WeakMap takes some non-zero amount
of time.

We’re going to go ahead and _try_ this approach and weight the costs
against the benefits — it’s easy to revert.