Prevent disconnects on simple re-mappings.
Conceptually, if you render `[foo, bar]` and then re-render with just
`[bar]` — you don’t expect that the `foo` element will be disconnected
and reconnected. However, if we naively first _move_ `bar` to the first
slot and then _delete_ `foo`… that‘s exactly what will happen.

Instead, we must first _delete_ `foo` and then _check_ to see if we
still need to _move_ `bar`.

Note that if a shuffle is _actually_ necessary, we cannot prevent a
disconnect / reconnect from occurring with the current DOM APIs.

Closes #253