Restrict usage of “style” tags in “x-parser”.
The goal here is to continue to reduce the surface area of this feature
to be as minimal as is sensible. In this case, basic browser APIs allow
authors to add style sheets to shadow roots (or documents) via the
`adoptedStyleSheets` setter. The most modern approach here is to write
an _actual_ `.css` file and import that as a CSS Module via the new(ish)
“import attributes” syntax.

This is much easier for browsers to optimize and simplifies our parsing
code a fair amount.

Closes #237