Remove “ifDefined” content guards.
Developers will no longer get errors thrown related to misused of the
“ifDefined” updater. This is not really a case that we need to guard
against though as it’s not ever mistakenly done in practice.

Note that the guard _does_ remain in place for now for boolean
attributes, defined attributes, and properties (which are common places
where mistakes occur). In the future, “ifDefined” will be deleted
completely and we’ll be able to do even more cleanup here.

Also, because we removed the “live” in a prior PR, we are also able to
move the change-by-reference detection guard up to the top-level
“commit” method for improved readability and performance.