Remove environmental fork from “x-parser” logic.
This change set moves _all_ DOM usage out of “x-parser” and into the
“x-template” file. Because “x-parser” is used both in Node and Browser
environments — we had previously needed to “mock” browser-only APIs for
our related “eslint-plugin-x-element” tool.

_Now_, the parser simply produces a stream of “tokens” as it interprets
interpolated html markup which can be consumed equally by the related
linter and by the default template engine.

It’s a critical design choice to make sure that the code performing the
linting is _byte for byte_ the _exact_ same as the code validating html
markup in the runtime — otherwise, we open up the door to unpredictable
and difficult-to-reproduce issues.†

Additionally, now that we don’t have forking logic in “x-parser”, we no
longer encounter Node / ESLint issues if `XParser` is a purely `static`
interface again. Therefore, this change set _also_ migrates us back to
a static interface such that integrators should not need to instantiate
an instance of the parser.

† “Yes” — A local Node version is not byte-for-byte the same as some
  local browser version, but our “x-parser.js” file will be the exact
  same in both places. That’s as much as we can control.