Restrict all JS-y escapes.
We previously allowed escaping the “\”, “$” and “`” characters to
accommodate for usage in template literals. However, this causes two
important forks (which has complexity and maintenance overhead):

 1. The “raw” and “cooked” strings cannot be assumed to be equivalent.
    This is a problem because linters that work with “raw” text in an
    IDE will need different indexing from outputs in the runtime (where
    text is “cooked”).
 2. Handling of escapes for certain characters differs from others. This
    is an issue because it increases our pattern matching logic which
    decreases performance and adds conceptual overhead to understand
    that _certain_ escaping was allowed, but _other_ escaping wasn’t.

The lowest-code fix for these conceptual forks is to simply _not allow_
any JS-y character escapes. That forces the “cooked” string to be
exactly the same as the “raw” string. It also greatly simplifies our
validation logic and gives us a small, but non-negligible perf win.

Closes #270.