Prepare codebase for new “moveBefore” DOM API.
Things seem to be _moving_ along pretty quickly for this new API. The
motivation for this (versus `insertBefore`) is that DOM state is not
reset when you _move_ DOM nodes around within the same connected tree.

Importantly, that means a `disconnectedCallback` followed immediately by
a `connectedCallback` will no longer happen. Those two lifecycle events
typically do some heavy lifting and we want to avoid them for cases like
nodes simply being moved around within the same list or something.

Note that this change set _prepares_ us to leverage this new API, but it
doesn’t yet commit to it as support is only currently in Chrome 133. We
wouldn’t normally get ahead of ourselves like this — but there are a lot
of positive signals from other browsers, so it seems like this will go
pretty quickly.

Final note — I was able to confirm that shuffling a keyed array won’t
cause disconnect / connect callbacks in the future!