Improve parser error messaging and factoring.
Taking on some of the language tweaks from #248. Also `switch`-ing on
the values (versus using a `Map`) so that these cases can be enforced by
code coverage.

We likely have a couple messages which are impossible to get to. It’d be
good to be forced to either remove them or find a way to prove that they
are reachable via our test suite.

For now, I reduced the coverage target to keep this change set smaller,
but I will address that in follow-up efforts.