Restore test coverage for “x-parser.js”.
In a prior change set, we changed our `Map` error mappings to `switch`
error mappings so that our code coverage tests would ensure that we are
able to prove that each error is possible to be shown — otherwise, why
keep them.

This change set was kept separate since it’s all about coverage.

Note that some of the regular expressions have changed — nearly all of
that is due to rewriting like `\s\n` as just `\s` since `\n` is one of
the characters matched by the `\s` set.