Stabilizes performance tests and adds screenshots.
The performance tests now run in a browser by default (when launched
from the command line). Additionally, you can save a screenshot of the
result for review.

This is particularly helpful when you want to perform and at-a-glance
check on whether a performance change being tested by an AI agent seems
to have meaningfully moved the needle or not.