Bump “eslint-plugin-jsdoc” for stricter linting.
This updates through some major versions of this dev dependency which
causes many stricter rules in the recommended set.

Mostly, they are all about declaring types as `unknown` versus `any`,
which is closer to what we _meant to say in the first place_.

The `any` type says — this is literally the union of everything that
exists. The `unknown` type says — we know literally nothing about this
thing.

Where the `any` type can be assigned to anything and anything can be
assigned to it. The `unknown` type cannot be assigned to anything, but
(importantly) anything can be assigned to it.

Closes #329.