Initialize properties in constructor for React.
React doesn’t provide a way for developers to signal that they want to
set a property (versus an attribute) in JSX — instead it performs some
magic to _guess_ what was meant. This _guess_ is based on whether the
expression `key in element` is `true`.

As a best-practice, XElement is _lazy_ about initialization and waits
until it is _actually connected to the DOM_ before initializing all
properties. But… that means React will always check for setters _before_
they exist!

This change set adds our accessor descriptors during construction, but
adds an internal flag to prevent certain behavior until the component is
fully initialized during connection.

Closes #331.