Adds experimental “decorators” interface.
The TC39 proposals for JavaScript decorators and decorator metadata have
made their way to stage 3. The next step is for browsers to begin
implementing. It is not an “if” anymore, it is a “when”.

This means that it’s time for us to explore what ergonomic improvements
might be unlocked by enabling interoperation with these new interfaces.

Some important notes:
- We could ship this now (only integrator code requires new syntax).
- The new interface _should_ be more idiomatic.
- We should be able to delete some invented concepts entirely.

Closes #346.