Use “moveBefore” for map shuffling.
Previously, both “connectedCallback” and “disconnectedCallback” would be
called when simply attempting to reshuffle elements in a mapping within
the same DOM tree.

With the addition of the “moveBefore” element API, we can fix this! Now,
such children are simply _moved_ and remain connected.

Additionally, a new test was added to confirm that browsers are calling
the “connectedMoveCallback” method on moves. And, another test added to
ensure that we are testing all conditions / ternaries in our code to
maintain 100% coverage.

Closes #254.