Disable JSR publishes.
We were only deploying to this registry to help us understand if it is
a good fit for the project or not. Currently, JSR is more of a
TypeScript-first project and we are a TypeScript-second project (i.e.,
we use JSDoc to type _actual JavaScript_ and then export `*.d.ts` files
from this library). Until they become more lenient on this front, it’s
less valuable to us anyhow!